+++ lib/himg.rb
@@ -8,7 +8,7 @@ require "himg/railtie" if defined?(Rails::Railtie)
 # Fall back to loading the non-versioned extension if version-specific loading fails.
 begin
   RUBY_VERSION =~ /(\d+\.\d+)/
-  require "#{Regexp.last_match(1)}/himg/himg"
+  require "himg/#{Regexp.last_match(1)}/himg"
 rescue LoadError
   require "himg/himg"
 end